Conflicts:
	pom.xml
	src/main/java/com/intetics/bean/EntitySchema.java
	src/main/java/com/intetics/controller/EntitySchemaController.java
	src/main/resources/application.properties
	src/main/webapp/WEB-INF/def/general.xml

+++ src/main/java/com/intetics/controller/EntitySchemaController.java
@@ -1,10 +1,15 @@
 package com.intetics.controller;
 
+import com.intetics.bean.Company;
 import com.intetics.bean.EntitySchema;
+import com.intetics.bean.User;
 import com.intetics.dao.EntitySchemaDao;
+import com.intetics.dao.UserDao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.util.Assert;
@@ -16,8 +21,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import javax.annotation.Nonnull;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
+import java.security.Principal;
 import java.util.Date;
 import java.util.List;
 
@@ -33,10 +37,14 @@ public class EntitySchemaController {
     @Autowired
     private EntitySchemaDao entitySchemaDao;
 
-    @RequestMapping(value = "/list")
-    public String getEntitySchemaList(ModelMap model) {
+    @Autowired
+    private UserDao userDao;
 
-        List<EntitySchema> entitySchemas = entitySchemaDao.getEntitySchemaList();
+    @RequestMapping(value = "/list")
+    public String getEntitySchemaList(ModelMap model, Principal principal) {
+
+        User user = userDao.getUserByEmail(principal.getName());
+        List<EntitySchema> entitySchemas = user.getCompany().getEntitySchemas();
         model.addAttribute("entitySchemaList", entitySchemas);
 
         LOGGER.trace("The list of entities has been requested");
@@ -45,13 +53,13 @@ public class EntitySchemaController {
     }
 
     @RequestMapping(value = "/create")
-    public String createEntitySchema(ModelMap model) {
+    public String createEntitySchema(ModelMap model, Principal principal) {
 
-        List<EntitySchema> entitySchemas = entitySchemaDao.getEntitySchemaList();
+        User user = userDao.getUserByEmail(principal.getName());
+        List<EntitySchema> entitySchemas = user.getCompany().getEntitySchemas();
         model.addAttribute("entitySchemaList", entitySchemas);
 
         EntitySchema entitySchema = new EntitySchema();
-        entitySchema.setName("");
         model.addAttribute("EntitySchema", entitySchema);
 
         model.addAttribute("modalTitle", "Create Entity");
@@ -63,16 +71,14 @@ public class EntitySchemaController {
     }
 
     @RequestMapping(value = "/edit/{entitySchemaId}", method = RequestMethod.GET)
-    public String startToEditEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model,
-                                          HttpServletRequest request) {
+    public String startToEditEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model) {
         Assert.notNull(entitySchemaId);
 
-        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+        EntitySchema entitySchema = verifyComplianceEntitySchemaAndCompany(entitySchemaId);
+        if(entitySchema == null)
+            return "error";
+
         model.addAttribute("EntitySchema", entitySchema);
-
-        HttpSession session = request.getSession();
-        session.setAttribute("entitySchemaCreateDate-" + entitySchemaId, entitySchema.getCreateDate());
-
         model.addAttribute("modalTitle", "Edit Entity");
         model.addAttribute("modalSaveButton", "Edit");
 
@@ -80,12 +86,17 @@ public class EntitySchemaController {
     }
 
     @RequestMapping(value = "/save", method = RequestMethod.POST)
-    public String saveEntitySchema(@ModelAttribute 
+    public String saveEntitySchema(@ModelAttribute
                                        @Validated(value = EntitySchema.MvcValidationSequence.class)
                                        EntitySchema entitySchema,
-                                   BindingResult bindingResult, ModelMap model, HttpServletRequest request) {
+                                   BindingResult bindingResult, ModelMap model, Principal principal) {
 
         Date currentDate = new Date();
+
+        User user = userDao.getUserByEmail(principal.getName());
+
+        Company company = user.getCompany();
+        entitySchema.setCompany(company);
 
         if (bindingResult.hasErrors()) {
             if (entitySchema.getId() == null) {
@@ -104,28 +115,35 @@ public class EntitySchemaController {
         }
 
         if(entitySchema.getId() != null){
-            HttpSession session = request.getSession();
-            entitySchema.setCreateDate((Date) session.getAttribute("entitySchemaCreateDate-"+entitySchema.getId()));
+            if(verifyComplianceEntitySchemaAndCompany(entitySchema.getId()) == null)
+                return "error";
+
             entitySchema.setModifiedDate(currentDate);
             entitySchemaDao.saveOrUpdate(entitySchema);
             return "redirect:/entity/" + entitySchema.getId() + "/field/list";
         }
 
-        entitySchema.setCreateDate(currentDate);
         entitySchema.setModifiedDate(currentDate);
+        entitySchema.setCompany(company);
 
         entitySchemaDao.saveOrUpdate(entitySchema);
         return "redirect:/entity/list";
     }
 
     @RequestMapping(value = "/delete/{entitySchemaId}/confirm", method = RequestMethod.GET)
-    public String startDeleteEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model) {
+    public String startDeleteEntitySchema(@Nonnull @PathVariable Long entitySchemaId, ModelMap model,
+                                          Principal principal) {
         Assert.notNull(entitySchemaId);
 
-        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+        EntitySchema entitySchema = verifyComplianceEntitySchemaAndCompany(entitySchemaId);
+        if(entitySchema == null)
+            return "error";
+
         model.addAttribute("EntitySchema", entitySchema);
 
-        List<EntitySchema> entitySchemas = entitySchemaDao.getEntitySchemaList();
+        User user = userDao.getUserByEmail(principal.getName());
+
+        List<EntitySchema> entitySchemas = user.getCompany().getEntitySchemas();
         model.addAttribute("entitySchemaList", entitySchemas);
 
         return "delete-entity";
@@ -135,9 +153,28 @@ public class EntitySchemaController {
     public String deleteEntitySchema(@Nonnull @PathVariable Long entitySchemaId) {
         Assert.notNull(entitySchemaId);
 
-        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+        EntitySchema entitySchema = verifyComplianceEntitySchemaAndCompany(entitySchemaId);
+        if(entitySchema == null)
+            return "error";
+
         entitySchemaDao.delete(entitySchema);
 
         return "redirect:/entity/list";
+    }
+
+    private EntitySchema verifyComplianceEntitySchemaAndCompany(Long entitySchemaId) {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User user = userDao.getUserByEmail(authentication.getName());
+
+        Company company = user.getCompany();
+
+        EntitySchema entitySchema = entitySchemaDao.getEntitySchema(entitySchemaId);
+
+        if(company.getEntitySchemas().contains(entitySchema))
+            return entitySchema;
+
+        return null;
     }
 }
